fix dynamic programming solving of MCKP in case there is a missing set number
The algorithm didn't work when the set number weren't all consecutive.
For example if item could either belong to set number 0 or set number 2 (there is no set number 1) then the algorithm didn't work.
This commit can't be cancelled in a real situation to improve the performance but in a pure MCKP problem with a lot of random number generation for tests i keep it.